Fixing tests broken due to copyright timestamp
All tests have failed since 2015-01-01 because the timestamp at the copyright notice of the generated files differs from the timestamp of reference file's copyright notice which is of 2014.

This is fixed by replacing all occurences of the current year in the generated files with "2014". Keep in mind that even new reference files need a timestamp of 2014 for the tests to work.